requirements.txt
setuptools
PyYAML
requests
kubernetes
datetime
configparser
slackclient
pyfiglet
redis
